Made DataScripts into local module/package. Improved DataScripts package structure. General maintanance
