+++ DataScripts/__init__.py
@@ -1,0 +1,4 @@
+# __all__ = ['data_manipulation', 'data_import', 'interpolate']
+from DataScripts.data_manipulation import manipulate_data
+from DataScripts.data_import import import_data
+from DataScripts.interpolate import interpolation
